+cleanup to make valgrind happy
git-svn-id: https://remotefs.svn.sourceforge.net/svnroot/remotefs/rfs_nss/trunk@1219 a487bacb-bc4e-0410-a857-d1bad4672f5a

+++ src/nss_cmd.h
@@ -45,13 +45,13 @@ enum nss_commands
 
 struct nss_command
 {
-	uint16_t command;
+	uint32_t command;
 	uint32_t data_len;
 };
 
 struct nss_answer
 {
-	uint16_t command;
+	uint32_t command;
 	uint32_t data_len;
 	int32_t  ret;
 	uint32_t ret_errno;
